Renamed POM files and fixes
