+++ e2e/pages/k8s-page.ts
